Shaders: split terrain.wgsl.ts by concern
The 836-line terrain shader module mixed three distinct
responsibilities. Split into:

  terrain-idw.wgsl.ts          IDW weighting and blending primitives
  terrain-containment.wgsl.ts  point-in-contour, signed distance, gradient
  terrain.wgsl.ts              high-level computeTerrainHeight/Normal

Shader code is byte-equivalent — every fn_* and struct_* code string is
moved verbatim. External importers all consume the high-level entry
points and need no change.

Also includes prettier formatting fix in boat/sail/README.md.
